Module Cohttp_lwt_unix_nossl.Server

The Server module implements the full UNIX HTTP server interface, including the UNIX-specific functions defined in S.

include Cohttp_lwt.S.Server with module IO = Cohttp_lwt_unix_nossl__Io

module IO : sig ... end

type conn = IO.conn * Cohttp.Connection.t

type response_action = [

| `Expert of Cohttp.Response.t * (IO.ic -> IO.oc -> unit Lwt.t)

| `Response of Cohttp.Response.t * Cohttp_lwt.Body.t

]

A request handler can respond in two ways:

Using `Response, with a Response.t and a Body.t.

Using `Expert, with a Response.t and an IO function that is expected to write the response body. The IO function has access to the underlying IO.ic and IO.oc, which allows writing a response body more efficiently, stream a response or to switch protocols entirely (e.g. websockets). Processing of pipelined requests continue after the unit Lwt.t is resolved. The connection can be closed by closing the IO.ic.

type t

val make_response_action : ?⁠conn_closed:(conn -> unit) -> callback:(conn -> Cohttp.Request.t -> Cohttp_lwt.Body.t -> response_action Lwt.t) -> unit -> t

val make_expert : ?⁠conn_closed:(conn -> unit) -> callback:(conn -> Cohttp.Request.t -> Cohttp_lwt.Body.t -> (Cohttp.Response.t * (IO.ic -> IO.oc -> unit Lwt.t)) Lwt.t) -> unit -> t

val make : ?⁠conn_closed:(conn -> unit) -> callback:(conn -> Cohttp.Request.t -> Cohttp_lwt.Body.t -> (Cohttp.Response.t * Cohttp_lwt.Body.t) Lwt.t) -> unit -> t

val resolve_local_file : docroot:string -> uri:Uri.t -> string

Resolve a URI and a docroot into a concrete local filename.

val respond : ?⁠headers:Cohttp.Header.t -> ?⁠flush:bool -> status:Cohttp.Code.status_code -> body:Cohttp_lwt.Body.t -> unit -> (Cohttp.Response.t * Cohttp_lwt.Body.t) Lwt.t

respond ?headers ?flush ~status ~body will respond to an HTTP request with the given status code and response body. If flush is true, then every response chunk will be flushed to the network rather than being buffered. flush is true by default. The transfer encoding will be detected from the body value and set to chunked encoding if it cannot be determined immediately. You can override the encoding by supplying an appropriate Content-length or Transfer-encoding in the headers parameter.

val respond_string : ?⁠flush:bool -> ?⁠headers:Cohttp.Header.t -> status:Cohttp.Code.status_code -> body:string -> unit -> (Cohttp.Response.t * Cohttp_lwt.Body.t) Lwt.t

val respond_error : ?⁠headers:Cohttp.Header.t -> ?⁠status:Cohttp.Code.status_code -> body:string -> unit -> (Cohttp.Response.t * Cohttp_lwt.Body.t) Lwt.t

val respond_redirect : ?⁠headers:Cohttp.Header.t -> uri:Uri.t -> unit -> (Cohttp.Response.t * Cohttp_lwt.Body.t) Lwt.t

val respond_need_auth : ?⁠headers:Cohttp.Header.t -> auth:Cohttp.Auth.challenge -> unit -> (Cohttp.Response.t * Cohttp_lwt.Body.t) Lwt.t

val respond_not_found : ?⁠uri:Uri.t -> unit -> (Cohttp.Response.t * Cohttp_lwt.Body.t) Lwt.t

val callback : t -> IO.conn -> IO.ic -> IO.oc -> unit Lwt.t

val resolve_file : docroot:string -> uri:Uri.t -> string

val respond_file : ?⁠headers:Cohttp.Header.t -> fname:string -> unit -> (Cohttp.Response.t * Cohttp_lwt.Body.t) Lwt.t

val create : ?⁠timeout:int -> ?⁠backlog:int -> ?⁠stop:unit Lwt.t -> ?⁠on_exn:(exn -> unit) -> 'cfg -> (_, 'flow) Conduit_lwt.protocol -> ('cfg, 't, 'flow) Conduit_lwt.Service.service -> t -> unit -> unit Lwt.t

create ?timeout ?backlog ?stop ?on_exn cfg protocol service t is a new HTTP server.

The user can decide to start a simple HTTP server (without encryption) or one with TLS encryption. It depends on what the user gives as cfg, protocol and service. Using conduit-lwt-tls, the end-user is able to make an encrypted HTTP server with:

let run =
  create cfg Conduit_lwt_tls.TCP.protocol Conduit_lwt_tls.TCP.service

A simple HTTP server (with conduit-lwt) is:

let run =
  create cfg Conduit_lwt.TCP.protocol Conduit_lwt.TCP.service

cfg depends on the given service - and let the user to define which port the server use, and, in the case of Conduit_lwt_tls.TCP.service, which TLS certificate it uses. See Conduit for more information about protocol and service.

When provided, the stop thread will terminate the server if it ever becomes determined.

When provided, backlog will limit the number of open connections.

Every connection will be served in a new lightweight thread that is invoked via the callback defined in t. If the callback raises an exception, it is passed to on_exn (by default, to a function that logs the exception using the Logs library).